refactor(editor): split reading block fields
